isdrawn function (#289)
implement an isdrawn function
slight fix to repetition detection
added "unit" tests
update ci

Bench: 2795524

+++ src/ucioptions.cpp
@@ -26,7 +26,7 @@ void uciOptions::printOptions()
 void uciOptions::uciHash(int value)
 {
     // value * 10^6 / 2^20
-    int sizeMiB = value * 1000000 / 1048576;
+    int sizeMiB = static_cast<uint64_t>(value) * 1000000 / 1048576;
     sizeMiB = std::clamp(sizeMiB, 1, MAXHASH);
     U64 elements = (static_cast<uint64_t>(sizeMiB) * 1024 * 1024) / sizeof(TEntry);
     TTable.allocateTT(elements);
